fix: refactor login code
